+++ src/app/(main)/main/_components/content/ContentList.tsx
@@ -1,6 +1,6 @@
-import { contentItemsMock } from "./mocks/contentItems";
 import ContentItem from "./ContentItem";
 import { useContentStore } from "@/stores/useContentStore";
+import { contentItemsMock } from "../mocks/contentItems";
 
 const ContentList = () => {
   const activeTab = useContentStore(state => state.activeTab);
