+++ src/app/(main)/main/_components/content/Platforms.tsx
